EY-4461 trekke ut funksjonalitet til PDFHelper
